requirements.txt
numpy
scikit-learn
statsmodels
matplotlib
pandas
pingouin
scipy
xlsxwriter
seaborn
progress
PyGithub
python-dotenv
scrapy
selenium
jupyter
nbconvert
prompt_toolkit
openpyxl
